requirements.txt
alembic==1.6.5
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
autopep8==1.5.7
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==8.0.1
cloudpickle==2.0.0
colorama==0.4.4
colorlover==0.3.0
cryptography==3.4.7
cufflinks==0.17.3
cycler==0.11.0
dateinfer==0.2.0
debugpy==1.6.0
decorator==5.1.1
decouple==0.0.7
defusedxml==0.7.1
Django==4.0.2
django-cors-headers==3.11.0
dnspython==2.1.0
email-validator==1.1.2
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.3
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
fonttools==4.33.1
greenlet==1.1.0
gunicorn==20.0.4
gym==0.23.1
gym-notices==0.0.6
idna==3.2
imbalanced-learn==0.9.0
imblearn==0.0
importlib-metadata==4.11.3
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.1
joblib==1.0.1
jsonpickle==2.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keras==2.8.0
kiwisolver==1.4.2
Mako==1.1.4
MarkupSafe==2.1.1
marshmallow==3.12.2
marshmallow-sqlalchemy==0.26.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mysqlclient==2.0.3
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
network==0.1
networkx==2.8
nltk==3.6.2
notebook==6.4.11
numpy==1.22.3
oauthlib==3.1.1
p5py==1.0.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pep517==0.12.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
plotly==5.7.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.3
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.11.2
PyJWT==2.1.0
pymongo==3.12.0
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyparsing==3.0.8
PyQt3D==5.15.5
PyQt5==5.15.6
PyQt5-sip==12.9.0
PyQtChart==5.15.5
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
pyrsistent==0.18.1
pyspark==3.2.1
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==0.18.0
python-editor==1.0.4
pytz==2021.1
pyvis==0.2.0
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
regex==2021.7.6
requests==2.26.0
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
six==1.15.0
sklearn==0.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.21
sqlparse==0.4.2
stack-data==0.2.0
statsmodels==0.13.2
tabulate==0.8.9
TBB==0.2
tenacity==8.0.1
tensorflow-metal==0.4.0
terminado==0.13.3
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.61.2
traitlets==5.1.1
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==3.6.0
WTForms==2.3.3
zipp==3.8.0
